+++ tcex/api/tc/ti_transform/ti_transform.py
@@ -20,7 +20,13 @@ class TiTransforms(TransformsABC):
         """Process the mapping."""
         self.transformed_collection: list[TiTransform] = []
         for ti_dict in self.ti_dicts:
-            self.transformed_collection.append(TiTransform(ti_dict, self.transforms))
+            self.transformed_collection.append(
+                TiTransform(
+                    ti_dict,
+                    self.transforms,
+                    seperate_batch_associations=self.seperate_batch_associations,
+                )
+            )
 
     @property
     def batch(self) -> dict:
@@ -57,6 +63,9 @@ class TiTransforms(TransformsABC):
                 continue
 
             # now that batch is called we can identify the ti type
+            if self.seperate_batch_associations:
+                associations = data.pop('association', [])
+                batch.setdefault('association', []).extend(associations)
             if isinstance(t.transform, GroupTransformModel):
                 batch['group'].append(data)
             elif isinstance(t.transform, IndicatorTransformModel):
@@ -71,6 +80,43 @@ class TiTransforms(TransformsABC):
 class TiTransform(TransformABC):
     """Threat Intelligence Transform Module"""
 
+    def add_custom_association(self, summary: str, indicator_type: str, association_type: str):
+        """Add a custom association."""
+        if not self.seperate_batch_associations:
+            raise TransformException(
+                field='associatedIndicator',
+                cause=RuntimeError(
+                    'Cannot add associated indicator to IndicatorTransformModel when '
+                    'seperate_batch_associations is False.'
+                ),
+                context=self.transformed_item,
+            )
+
+        self.transformed_item.setdefault('association', []).append(
+            {
+                'ref_1': self.transformed_item['summary'],
+                'type_1': self.transformed_item['type'],
+                'ref_2': summary,
+                'type_2': indicator_type,
+                'association_type': association_type,
+            }
+        )
+
+    def add_associated_indicator(self, summary: str, indicator_type: str):
+        """Add an associated indicator."""
+        if not self.seperate_batch_associations:
+            self.transformed_item.setdefault('associatedIndicators', []).append(
+                {'summary': summary, 'indicatorType': indicator_type}
+            )
+        else:
+            self.transformed_item.setdefault('association', []).append(
+                {
+                    'ref_1': self.transformed_item['xid'],
+                    'ref_2': summary,
+                    'type_2': indicator_type,
+                }
+            )
+
     def add_associated_group(self, group_xid: str):
         """Add an associated group.
 
@@ -79,12 +125,38 @@ class TiTransform(TransformABC):
                 {
                     'groupXid': 'dd78f2b94ac61d3e5a55c1223a7635db00cd0aaa8aba26c5306e36dd6c1662ee'}
         """
-        # process type specific data
-        if isinstance(self.transform, GroupTransformModel):
-            self.transformed_item.setdefault('associatedGroupXid', []).append(group_xid)
+        if not self.seperate_batch_associations:
+            # process type specific data
+            if isinstance(self.transform, GroupTransformModel):
+                self.transformed_item.setdefault('associatedGroupXid', []).append(group_xid)
+            elif isinstance(self.transform, IndicatorTransformModel):
+                associated_group = {'groupXid': group_xid}
+                self.transformed_item.setdefault('associatedGroups', []).append(associated_group)
+        elif isinstance(self.transform, GroupTransformModel):
+            self.transformed_item.setdefault('association', []).append(
+                {
+                    'ref_1': group_xid,
+                    'ref_2': self.transformed_item['xid'],
+                }
+            )
+            self.log.info(
+                'Added associated group with xid: %s to %s',
+                group_xid,
+                self.transformed_item['xid'],
+            )
         elif isinstance(self.transform, IndicatorTransformModel):
-            associated_group = {'groupXid': group_xid}
-            self.transformed_item.setdefault('associatedGroups', []).append(associated_group)
+            self.transformed_item.setdefault('association', []).append(
+                {
+                    'ref_1': group_xid,
+                    'ref_2': self.transformed_item['summary'],
+                    'type_2': self.transformed_item['type'],
+                }
+            )
+            self.log.info(
+                'Added associated group with xid: %s to %s',
+                group_xid,
+                self.transformed_item['summary'],
+            )
 
     def add_attribute(
         self,