+++ release_notes.md
@@ -4,6 +4,7 @@
 
 -   APP-4851 - [Util] Updated regex patterns for Playbook and TC variable to be more specific
 -   APP-4852 - [API] Send query params in body for GET requests
+-   APP-4846 - [transform] Add support for Batch API update which enables top-level associations.
 
 
 ## 4.0.8